feat: Refactor consumer management and user levels integration
- Removed consumer deletion functionality from ConsumerTable component.
- Updated MCPFormDrawer to handle allowed consumer levels instead of consumers.
- Modified MCPDetailPage to reflect changes in consumer authorization handling.
- Enhanced RouteForm to utilize user levels for authorization configuration.
- Introduced new utility functions for user level normalization and expansion.
- Updated services to manage invite codes and consumer password resets.
- Added new APIs for managing user policies in AI quota.
- Improved Helm chart versioning and deployment scripts.
- Documented architectural changes and migration strategies in memory and project files.
- Added TODOs for future enhancements related to consumer integration.

+++ backend/console/src/main/java/com/alibaba/higress/console/model/aiquota/AiQuotaConsumerQuota.java
@@ -17,5 +17,5 @@ public class AiQuotaConsumerQuota {
     private String consumerName;
 
     @Schema(description = "Current quota")
-    private int quota;
+    private long quota;
 }
